Implement card on project & algorithms
