+++ .gitignore
@@ -10,7 +10,6 @@
 
 # dependencies
 /node_modules
-node_modules
 
 # profiling files
 chrome-profiler-events.json
