+++ projects/components/src/lib/hyt-detailed-card/hyt-detailed-card.component.scss
@@ -62,7 +62,7 @@ $primaryDarker: #094894;
             background-repeat: no-repeat;
             background-position: bottom center;
             background-size: cover;
-            transition: all 0.15s ease-in;
+            transition: all .15s ease-in;
             transform: translateY(100%) scale(0);
             opacity: 0;
