--- src/app/pages/algorithms/algorithm-card/algorithm-card.component.html
@@ -1,41 +1,0 @@
-
-<mat-card class="mat-elevation-z5">
-    
-    <div class="title-card-back">
-        {{ algorithm.name }}
-    </div>
-
-    <div class="box-card-front">
-
-        <div class="header-card-front">
-
-            {{ algorithm.name }}
-
-        </div> <!-- End Card Front Header -->
-
-        <div class="content-card-front">
-
-            <div class="card-description">
-                {{ algorithm.description }}
-            </div> <!-- End Card Description -->
-
-        </div> <!-- End Card Front Content -->
-
-        <div class="footer-card-front">
-
-            <hyt-button
-                class="view-algorithm-btn" 
-                routerLink="/algorithm-wizard/{{algorithm.id}}"
-            >
-                View Algorithm
-            </hyt-button>
-            
-        </div> <!-- End Card Front Footer -->
-
-    </div>
-
-</mat-card>
-
-<div class="delete-algorithm-btn" (click)="deleteAlgorithm()">
-    <span class="hyt-icon icon-hyt_cancel"></span>
-</div>
