+++ src/styles.scss
@@ -13,23 +13,29 @@
         clip-rule: evenodd;
         fill: #144b64;
     }
+
     .blueDark {
         fill: #001a24;
     }
+
     .Glow {
         fill: #001a24;
         filter: url(#glow);
     }
+
     .st0 {
         filter: url(#glow);
     }
+
     .st1 {
         fill: #001a24;
     }
+
     .st2 {
         opacity: 0;
         fill: none;
     }
+
     .st3 {
         opacity: 0.6;
         fill: #ea0707;
@@ -52,11 +58,46 @@
     font-size: 0.8rem;
     background-color: #266dca;
 }
-.hyt-card-container {
-    display: flex;
-    flex-wrap: wrap;
-    margin-bottom: 1rem;
-    @media screen and (max-width: 1153px) {
-        justify-content: center;
+
+.hyt-card-wrapper {
+    width: 100%;
+    margin: 0 auto 1rem;
+
+    .hyt-card-container {
+        display: flex;
+        flex-wrap: wrap;
+        justify-content: flex-start;
+        width: 100%;
+
+        >* {
+            flex: 1;
+            margin-right: auto;
+        }
+
+        
     }
+
+    .hyt-card-container:after {
+        content: '';
+        flex: auto;
+    }
+}
+
+@media screen and (min-width: 1154px) {
+    .hyt-card-wrapper{
+        width: 90%;
+        max-width: 90%;
+    }
+}
+@media screen and (max-width: 550px) {
+    .hyt-card-wrapper{
+        .hyt-card-container {
+            justify-content: center;
+    
+            >* {
+                flex: none;
+                margin-left: auto;
+            }
+        }
+    } 
 }
